+++ QuickLook.Plugin/QuickLook.Plugin.MarkdownViewer/Plugin.cs
@@ -59,6 +59,7 @@ public sealed class Plugin : IViewer
         return AsciiDocPanel.CanHandle(path)
             || IpynbPanel.CanHandle(path)
             || MermaidPanel.CanHandle(path)
+            || RstPanel.CanHandle(path)
             || _markdownExtensions.Any(path.ToLower().EndsWith);
     }
 
@@ -85,6 +86,12 @@ public sealed class Plugin : IViewer
         {
             var panel = new MermaidPanel();
             panel.PreviewMermaid(path);
+            _panel = panel;
+        }
+        else if (RstPanel.CanHandle(path))
+        {
+            var panel = new RstPanel();
+            panel.PreviewRst(path);
             _panel = panel;
         }
         else