+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Add FilePilot preview integration by @Andrey Semjonov [#1949](https://github.com/QL-Win/QuickLook/issues/1949)
 - Add AsciiDoc support to markdown viewer (`.adoc`, `.asciidoc`, `.asc`, `.ad`)
 - Add Jupyter Notebook support to markdown viewer (`.ipynb`)
+- Add reStructuredText support to markdown viewer (`.rst`, `.restructuredtext`)
 - Add configurable keyboard shortcuts to toggle TOC visibility for markdown viewer [#1934](https://github.com/QL-Win/QuickLook/issues/1934)
 - Add option to disable window show transitions `<ShowWindowTransition>False</ShowWindowTransition>`
 - Add `.winmd` to PE viewer supported extensions